Now that works
I felt this appropriate because Bill Maher is a socialist and while
Obama may not be, both offer an interesting discourse in this full
interview.

I am a fan of Noam Chomsky and feel that he offers a level of deep
thought that many do not or cannot offer. He was crowned by Google as
the “Snowden without the travel restriction.”